requirements.txt
opencv-python~=4.8.0.74
pytesseract~=0.3.10
streamlit~=1.25.0
numpy~=1.25.2
Pillow~=9.5.0